bblogger: modify type of angle step to float
This triggered a build fail, and it's good it did. The calculated
angle step surely must be float, although for better results a
nice number of steps should be chosen.

+++ src/plugins/laserht/sensproc_thread.h
@@ -90,7 +90,7 @@ class LaserHtSensorProcThread
   HoughTransform *__ht;
   unsigned int    __num_vals;
   int           **__values;
-  int             __angle_step;
+  float           __angle_step;
   float           __r_scale;
 };
 #endif
